telegram_bot/requirements.txt
aiogram
yt-dlp
openai-whisper
deep-translator
ffmpeg-python